backend/requirements.txt
Flask
Flask-Cors
python-dotenv
azure-ai-textanalytics
PyPDF2
openai
requests
beautifulsoup4
python-docx
openpyxl